Fix erroneous `mux_direction == 'dut'`
mux_direction is only ever 'dut_sees_usbkey' or 'servo_sees_usbkey'.
However, the function checks to see whether mux_direction == 'dut'.
That will never be true. So, this CL changes it to check whether
mux_direction == 'dut_sees_usbkey'.

The code modified here was added in the following CL:
https://crrev.com/c/2073517
It looks like this issue was noticed and commented on after the CL was
merged, but was never fully addressed.

BUG=None
TEST=None

Change-Id: I44ecf925c38a0dfe0978450f7dadcf57def2ac85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2216409
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>