Let firmware_install() raise TestError if package download fails
firmware_install() is quite a long method, and it used to take
time to find a failure point if any happens. This patch let
firmware_install() raise a TestError with proper text if it catches
any exceptions during firmware download, with a proper message.

BUG=none
TEST=none

Change-Id: I213c31963eb63c8c3d6f00b4dd5686b5c0cecf4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2220775
Tested-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Namyoon Woo <namyoon@chromium.org>