platform_PrinterPpds: Skip modules loading and mounting / as read-only
This patch removes code responsible for loading and unloading modules
with filters for Epson and Star printers. These filters are now
included in the image. This patch also removes remounting the root
partition in read-only mode to solve the bug mentioned below. It only
matters for "debug" runs, when some temporary changes must be made to
CUPS configuration file and the root partition is remounted in
read-write mode and then remounted back in read-only mode. It may cause
an error (like "partition is still in use"). The "debug" mode is run
only manually by developers, so remounting the root partition back as
read-only is not necessary.

BUG=chromium:1070023
TEST=on atlas

Change-Id: Ibf902a998c2d3e34096114e99a4c13923c4d12a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2220514
Reviewed-by: Sean Kau <skau@chromium.org>
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Auto-Submit: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>