wifi: fix get_disconnect_reasons with non-ASCII characters
When devices crash log files often end up with non-ASCII characters in
some parts. When grep'ing for specific messages in the logs, we must
tell grep to always handle the log files as text files. Otherwise grep
will assume that the logs are binary if it encounters a non-ASCII
character anywhere and will fail.

BUG=b:157775548
BUG=chromium:812685
BUG=chromium:698859
TEST=network_WiFi_DisconnectReason.switch_ap went from failing to
passing on a DUT that had non-ASCII characters in /var/log/net.log

Change-Id: I6d7a755f20005476fac4325b1111009955cb31c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2222888
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>