Fix guest mode RLZ Data file check
The test was unable to find the file /home/chronos/user/'RLZ Data' file
while running in guest mode, due to the new way guest sessions are
started. The guest user mount is no longer in the root mount namespace,
so the command to read the file needed to be modified to check in the
new namespace.

BUG=chromium:1048226
TEST=test_that chromeos6-row4-rack12-host21.cros rlz_CheckPing.guest --autotest_dir ~/trunk/src/third_party/autotest/files -b eve
and also with -i 'eve-release/R84-13045.0.0', a version where the guest
mount change was reverted, just in case it's ever reverted again for
some reason.

Change-Id: Ifcab416c0134da6a91f5e12ecba02ca19d39a6b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2224780
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>