network_WiFi_SimpleConnect: remove checkNonAsciiSSID test
Currently the testcase does not properly verify the behavior as the
SSIDs to test are not assigned into HostapConfig in control file.

Also, the framework can not handle these SSIDs now because:
- need to use printf escaped string e.g. ssid2=P"ab\n" for arbitrary
  binary SSIDs in hostapd.conf, but the logic is not yet available in
  Autotest.
- With non ascii printable characters, the WiFi service name will not
  match SSID. (non ascii printable characters are replaced with '?')
  We'll need to find the service with WiFi.HexSSID instead.

As one similar test is already added in Tast in crrev.com/c/2200522,
let's remove this one from Autotest and prefer the one in Tast.

BUG=b:156313361
TEST=None

Change-Id: Id63e0271f413b1b52562d51829d4485dcb23973a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2224685
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>