autotest: audio_CrasDevSwitchStress: Restore the audio nodes
This patch saves and restores the audio nodes in order to get back
the state in which the test started. Since, the test switches between
the nodes, it could leave the dut in a state which could cause
subsequent test to fail.

BUG=b:158058187
TEST=test_that <ip> audio_CrasDevSwitchStress.input audio_AlsaLoopback

Change-Id: I0f00a0001e5c8ee4a2b3c0bdd2f1ca66d2d32ab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2233940
Tested-by: Akshu Agrawal <akshu.agrawal@amd.com>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Auto-Submit: Akshu Agrawal <akshu.agrawal@amd.com>