FAFT: Add telemetry for USB disk related issues
There are a couple of FAFT failures related to FAFT not finding the USB
storage. It finds no device at all, or doesn't find the partition or
can't mount the file system.

On such an error, collect some more data on what's going on.

BUG=b:155774764
TEST=local modifications to force the exception in both places led
to logs that contained useful information.

Change-Id: Iab87f413693e3b8f880398f1dc102a61bd91ecdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2238286
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>