Add SkiaRenderer variations to graphics_HwOverlays
This CL creates 3 variations of the graphics_HwOverlays test which force
it to use the SkiaRenderer instead of the default (GLRenderer). This is
intended to aid during development to ensure overlays don't regress
during the transition to SkiaRenderer.

These variations will be removed when SkiaRenderer is enabled by default
on Chrome OS (https://crbug.com/925181).

BUG=b:159343247
TEST=fails on nocturne with "number of overlays is 1" (as expected)

Change-Id: Ib3a657c4705f0c661b7cf9156d9d85404e68fec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2253125
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>