firmware_Fingerprint: Remove old firmware versions
We don't need to keep the old firmware versions listed in the test; we
can see the version history from both the git repo where the firmware is
stored as well as the git history of this test file. The list was
starting to grow large. Also, we don't want to get into a situation
where an old (valid) version of the firmware is added to the build and
the tests still pass.

BUG=none
TEST=none

Change-Id: I119f0331f59331f70a41d671ebf96cedb87412b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2307715
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Auto-Submit: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Yicheng Li <yichengli@chromium.org>