autoupdate_EndToEndTest: Use quick-provision for installing the source image
Currently, autoupdate_EndToEndTests use AU to install very old images in
order to test AU success on old images. This means that newer
update_engines should be able to apply a very old update payload. This
has caused us issues in many places as it has hindered us to deprecated
old features and improve update_engine to a more efficient
manner. Now that we have quick provision payloads in release buckets, we
can use them to install the source image of end to end test. This CL
does that.

However, the only caveat here is that the fallback code to provision
with AU when the QP fails does not work because the release buckets do
not have the test AU payloads in them. So we added a flag (au_fallback)
and disabled that fallback. And we added a temporary fallback to the
original provision mechanism using the AU payloads. We need to remove
this once the EndToEnd tests are stable and showing no sign of failures
in quick-provision.

BUG=chromium:991421
TEST=test_that --autotest_dir ~/trunk/src/third_party/autotest/files/ --args="target_release=13280.0.0 target_payload_uri='gs://chromeos-releases/dev-channel/reef/13290.0.0/payloads/chromeos_13290.0.0_reef_dev-channel_full_test.bin-gvswkobwg43gf4byzyw75rj4io3sl5y5' source_release=13280.0.0 source_payload_uri='gs://chromeos-releases/dev-channel/reef/13290.0.0/payloads/chromeos_13290.0.0_reef_dev-channel_full_test.bin-gvswkobwg43gf4byzyw75rj4io3sl5y5' update_type=full" chromeos6-row4-rack10-host12.cros.corp.google.com autoupdate_EndToEndTest
TEST=above test after raising an exception in
_quick_provision_with_gs_cache() to make sure
_quick_provision_with_devserver() works too.

Change-Id: I1ac1828f58c9a63f4a7d8888e651248033d567ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2261239
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>