Delete site_test_importer.sh
ChromeOS has a P1 Q3 OKR to remove oppressive language from our code,
including renaming "whitelist" to "allowlist".
My original intent with this CL was to rename site_whitelist_tests.

But it seems that site_whitelist_tests was only used in
site_test_importer.sh, which appears to be unused. In fact, that script
relies on test_importer.py, which was deleted in
https://crrev.com/c/1967065.  So, let's just get rid of the dead code.

BUG=b:159917945
TEST=None

Change-Id: I6bfee90277572fddba30c4f17e4eec4cac765cb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2268284
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>