autotest: relax graphics_VideoRenderingPower CPU idle threshold
graphics_videoRenderingPower waits for up to 60s for the
CPU utilization % to go under 0.2%. This seems challenging
for many devices, e.g. octopus, so this CL relaxes that
threshold to 2.5%. The CPU idling is not part of the test,
but a way to separate the measurements associated to the
different test cases inside, so this numbers should not
affect the outcome.

Bug: chromium:1006882
Change-Id: I9dec0c343defe5fc6df3e064cdfa040c9614cb4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2274592
Tested-by: Miguel Casas <mcasas@chromium.org>
Auto-Submit: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>