autotest: Move host dependency check inside verifier.
Moved it to minimize fail case if host is not available.
AdminAudit is starting with the set of actions and all of them has to
run if possible. By this move we allowed each verifier to check what
required to run.
If dependency not provided we can skip of the action.

BUG=chromium:1069101, chromium:1059439
TEST=unittests, presubmit, run local

./site_utils/admin_audit/main.py --hostname chromeos1-row4-rack4-host4   --results-dir /tr --host-info-file /tr/host_info_store/chromeos1-row4-rack4-host4.store   verify-dut-storage verify-servo-usb-drive verify-servo-fw
./site_utils/admin_audit/main.py --hostname chromeos1-row4-rack4-host3   --results-dir /tr --host-info-file /tr/host_info_store/chromeos1-row4-rack4-host3.store   verify-dut-storage verify-servo-usb-drive verify-servo-fw

Change-Id: I9dab24a61290970ff95b4e0431c6f74381f972cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2253052
Tested-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Otabek Kasimov <otabek@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>