autotest: make tests less fragile and more robust to uprev process
Some existing tests enumerate subsets of boards by name. This is inefficient
and toilsome with uprevs in this way: you see a test regression on the uprev,
spend time digging in to it and find that the fix is to change the test to
include $BOARD-kernelnext along side $BOARD. Test logic can be written to
expect and optional "-kernelnext" and cut out all this toil and inefficiency.

BUG=b:160605128
TEST=manual, ran test

Change-Id: Idfe62549d3c618ff9ae3d83200a159dde2eb34dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2283685
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Bob Brandt <brndt@google.com>
Tested-by: Bob Brandt <brndt@google.com>
Auto-Submit: Bob Brandt <brndt@google.com>