firmware_Cr50CCDFirmwareUpdate: set 'TIME' attribute as 'FAST'
This patch changes 'TIME' attribute of firmware_Cr50CCDFirmwareUpdate as
'FAST' so that the autotest can schedule it later than other tests
attributed as longer TIME. The cleanup from firmware_Cr50CCDFirmwareUpdate
sometimes fail to recover a dut firmware, and it affects the rest of the test.
This is to minimize this side-effect as much as possible.

BUG=None
TEST=ran 'test_that --pretend suite:faft_cr50_prepvt' and checked its
sequenc order changed from the 5th to 31st among thirty six tests.

Change-Id: I480fec99049b1ed88c761fcf40afe27ccf77503a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2289014
Tested-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>