autoupdate_EndToEndTest: Use quick-provision for installing the source image
Currently, autoupdate_EndToEndTests use AU to install very old images in
order to test AU success on old images. This means that newer
update_engines should be able to apply a very old update payload. This
has caused us issues in many places as it has hindered us to deprecated
old features and improve update_engine to a more efficient
manner. Now that we have quick provision payloads in release buckets, we
can use them to install the source image of end to end test. This CL
does that.

This is the second attempt for this feature. The first one was
crrev.com/c/2261239 which was eventually reverted in
crrev.com/c/2288160. The difference is in this provisioning we don't
update the device hostinfo labels anymore so we don't get the problem in
crbug.com/1103324. In old style of this test we didn't update the
hostinfo lables, so no need to do it now.

This CL only gets applied for an example board 'reef'. This allows us to
make sure the CL is sane before releasing it on all channels. Once we
know the process works for 'reef' board, we can release this for all boards.

BUG=chromium:991421
TEST=test_that --autotest_dir ~/trunk/src/third_party/autotest/files/ --args="target_release=13280.0.0 target_payload_uri='gs://chromeos-releases/dev-channel/reef/13290.0.0/payloads/chromeos_13290.0.0_reef_dev-channel_full_test.bin-gvswkobwg43gf4byzyw75rj4io3sl5y5' source_release=13280.0.0 source_payload_uri='gs://chromeos-releases/dev-channel/reef/13290.0.0/payloads/chromeos_13290.0.0_reef_dev-channel_full_test.bin-gvswkobwg43gf4byzyw75rj4io3sl5y5' update_type=full" chromeos6-row4-rack10-host12.cros.corp.google.com autoupdate_EndToEndTest
TEST=above test after raising an exception in
_quick_provision_with_gs_cache() to make sure
_quick_provision_with_devserver() works too.

Change-Id: I9090d9d3ed10eafc1907055885a44e9174eb3262
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2289886
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>