ssh_host: Rename run method so it's easy to find
When trying to find implementations of the base run() method, the
run_very_slowly method doesn't come up.  Now it's just named run,
but decorated with a note that it's slow.  The type of self._client
is also now declared, so it knows which run() method to choose.

BUG=None
TEST=Run firmware_FAFTRPC.all and utils/unittest_suite.py

Change-Id: I5922b2b34828e9cd89399eb7466d72c3c8744901
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2290572
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>