bluetooth: TestNA shouldn't fail batch
If a test fails because there aren't enough peers, it shouldn't prevent
the rest of the batch from running. Move the check for number of peers
into the try block so the TestNA is considered SKIPPED.

Also, move the skip_models check before the peer initialization so that
tests run faster if they're being skipped.

BUG=None
TEST=Ran bluetooth_AdapterSRSanity with only 1 peer and verified entire
     batch ran. On Zork, verified peer wake tests were skipped without
     initializing the peer.

Change-Id: I2a892b5107cb63130bbfa58fd50428115378c500
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2291600
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Alain Michaud <alainm@chromium.org>