cheets_CTS_P: Enable 64-bit nativebridge flag on dev channel.
The definition of "dev channel" here is that the branch number
("z" of Rxx-yyyyy.z.w) is zero. The flag enables the feature
only for the boards built with the corresponding USE flag.
For other borads, this is no-op.

BUG=b:161002488
TEST=cheets_CTS_P.9.0_r12.arm.CtsUsbTests on fizz runs 2 tests

Change-Id: Ic24de5fbf69af8ecd7e4771e260a8574fdaae295
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2291428
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Jaydeep Mehta <jaydeepmehta@google.com>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>