pd_faft: Allow an empty string in the PE state
When the Type-C state is not in an attached state, the policy engine
(PE) is not running that the PE state returns an empty string. The
regular expressions should allow this case.

BUG=b:160940817
TEST=Manually tested the modified regular expressions, which can handle
the empty PE state and the normal cases.

Change-Id: Ida0ba8a0dcc3b13c9a1154a967d6a9289cae2bde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2295255
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Sam Hurst <shurst@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Auto-Submit: Wai-Hong Tam <waihong@google.com>