Detect inheritance loops in fw-testing-configs
If malformed JSON is ever uploaded to fw-testing-configs containing an
inheritance loop (e.g. A.parent = B; B.parent = A), that would be
impossible to resolve. Prior to this CL, the error-state would be an
infinite loop, which would be difficult to debug.

With this CL, infinite loops are detected, and throw a descriptive
error, so it will be easier to debug a failure of this type.

Also added a unit test for this new behavior.

BUG=b:160984219
TEST=config_unittest.py
TEST=firmware_FAFTSetup

Change-Id: I20698767b4104f35b04b620bd282c5ba83ef703b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2296493
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>