Reland "Revert "provision_AutoUpdate: Do not stage AU payloads unless necessary""
This reverts CL crrev.com/c/2296498.

Currently provisionning always stages the AU full payloads at the
beginning. But majority of provision runs succeed by quick-provisioning
and they never get to AU provisioning. So this is a waste of time and
space trying to stage large files that are not going to be used. This CL
fixes that problem.

This CL fixes the problem discussed in crbug.com/1105475 by staging the
autoupdate files for chromite's ChromiumOSUpdater too.

BUG=chromium:1097995
TEST=test_that --args="value='reef-release/R85-13280.0.0'" chromeos6-row4-rack10-host19.cros.corp.google.com provision_AutoUpdate
TEST=test_that --args="value='reef-release/R85-13280.0.0'" chromeos6-row4-rack10-host19.cros.corp.google.com provision_AutoUpdate.double

Change-Id: Iee2d3158dbac90d310eee736c4553a2b423e4bac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2300642
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>