autotest: Fixes policy_AlternateErrorPages
The autotest for policy_AlternateErrorPages was broken because the
behaviour changed so that for a request for localhost:8080 the same
error with no different error pages would be shown.

It the request is changed to localhost-8080 the different error pages
dependent on the policy setting can be observed.

This CL implements this fix.

BUG=b:161347481
TEST=Run test on applicable DUT (kvm)

Change-Id: Ie9dbfa1b7bcb5c4cb5e4178a5ed93d43f05f4f47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2300299
Tested-by: Marc Grimme <marcgrimme@chromium.org>
Commit-Queue: Aya Elsayed <ayaelattar@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>