tast: Support downloading data files lazily.
This change adds download_data_lazily optional parameter to tast.py.
If it is True, -downloaddata=lazy is passed to the tast command,
which instructs it to download external data files lazily between
tests rather than downloading them in a batch before tests.

The new parameter is False by default, and at this point no control
file uses it. It is turned on for several control files in later
changes.

BUG=chromium:965703
TEST=tast_unittest.py

Change-Id: Iaaf562ae796acab5574302312b3f4a2040664e90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2321813
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>