FAFT: Added new power button check for devices without display
Added an additional power button sequence check for devices without a
display. When no display is detected the power button is pressed for a
short 200ms duration and the test checks power state to verify the DUT
shut down (the intended behavior for chromeboxes).

BUG=b:159944769
TEST=Verified test passed for device with display (kohaku) and without
     display (teemo)

Change-Id: I4de4d4461b8ca7c7942d662f534438f30b9b87af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2308940
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Brent Peterson <brentpeterson@chromium.org>