autotest: Fixes policy_ForceYouTubeRestrict
YouTube seems to have changed their UI in terms of how they handle
restricted mode. This caused this test to break.

This fix includes incorporating the new UI into the test. Some code
that tested youtube restrictions has been removed in favour to test
the policy behaviour.

There are still some glitches like the settings menu in YouTube doesn't
reflect on the policy setting but just shows a message on the webpage.
As soon as this is fixed on the YouTube side it can be incorporated
again. Currently this part is commented out but the code should be
there.

BUG=b:161448116
TEST=Run test on applicable DUT (kvm)

Change-Id: I2f314cb29b7bbeba22dc8be0fcc93244d8b1aa3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2302902
Reviewed-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Marc Grimme <marcgrimme@chromium.org>
Tested-by: Marc Grimme <marcgrimme@chromium.org>
Auto-Submit: Marc Grimme <marcgrimme@chromium.org>
Commit-Queue: Aya Elsayed <ayaelattar@google.com>