Tauto: Use Servo for some 'tast' control files
Some Tast tests require the use of Servo. When the Tauto 'tast' test is
run, before this CL no servohost was instantiated, so Tast tests
requiring Servo would fail. Because of this, Tast tests which use Servo
fail when run in the CQ: for example, tastboard/test/example.ServoEcho.

With this change, the Autotest "Tast" test initializes servod on the
servo host for a few control files:
* critical-system, which runs in the CQ
* informational-system, which runs in the CQ
* firmware-cr50, because many firmware tests require servo
* generic-servo, a new clone of the existing 'generic' control

When these tests are run without the servo_host and servo_port
command-line args, they may now fail to connect to the servo on the
default host/port (localhost:9999), and so raise an error.

BUG=b:156771613
BUG=chromium:1087369
TEST=Run generic-servo with tast_expr='example.ServoEcho', check success
TEST=Run all other affected controls, check that they run

Change-Id: I7043dfb63f99779e272fe1a276fd8e536cf90d6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2304912
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>