firmware_ECBootTime: Enable port80 output on x86
The EC boot time test is relying on a signal from the host that it
is booting.  Currently the way things are configured this is only
noticed when the host issues the first transaction to the EC.

On x86 we used to have port80 output that the test also looks for,
but that was squelched in recent platforms because it would
interrupt the EC hash calulcation and delay the boot.

The test can re-enable this output however and get an early signal
from the host that the SOC is alive and booting.

BUG=b:160540281
TEST=pass firmware_ECBootTime on volteer

Change-Id: I0b21422e9b9366c86c4d19593910b71e4c7b2863
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2309069
Reviewed-by: Alex Levin <levinale@google.com>
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Dossym Nurmukhanov <dossym@chromium.org>