firmware_ECCharging: Cache battery information
We send `battery` EC commands multiple times during this test to probe
different battery parameters.

This is not efficient - we only need to run `battery` command once
and then collect all of the battery parameters we are interested.

Besides, when the `battery` commands are sent consecutively in a short
time, we occasionally observe character lost in EC UART output, which
makes the test fail.

To make the test more efficient and robust, let's fix it.

BUG=b:160186447
TEST=run the test on a blorb DUT, pass 10 times in a row

Change-Id: Idf5fdd3aebf556957827fa2da65e7cd3e3952caf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2311658
Commit-Queue: Philip Chen <philipchen@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>