firmware_ECUsbPorts:
Change the test status to TEST_NA when using a servo v4 type-A and the
dut only supports a single USB-A control. With this combination,
disabling the USB-A port also disconnects the network connection from
the DUT causing a test failure.

BUG=b:159450224
TEST=Verify test passes on Volteer with servo_v4 type-C
TEST=Verify TEST_NA result on Volteer with servo_v4 type-A

Change-Id: Ib5858a24cb2765bfa74f06cae69cf9270d9d9764
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2316713
Tested-by: Keith Short <keithshort@chromium.org>
Auto-Submit: Keith Short <keithshort@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Abe Levkoy <alevkoy@chromium.org>