power_MeetClient: Workaound for Chrome telemetry
On some dut, camera won't load first on the first
time when using meet with '--use-fake-ui-for-media-stream'
chrome option. This does not happen with manually
test using normal user flow.

While the root cause is still unknown, I found that
reload the page resolve the issue. This CL adds
code to do that as a workaround for the test.

BUG=b:153445204
TEST=powe_MeetCall.4bot on Eve

Change-Id: I126fb425fb153bc29a39dfe790fa21bf78f50ac2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2316718
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>