autotest: network_WiFi_RoamFT: Use D-Bus roam command instead
The test previously exploited wpa_supplicant's roam algorithm,
particularly its preference for 5GHz APs, to test a roam. There's no
need for this since the purpose of this test is simply to ensure that
we're able to do an FT roam from one AP to another. Replace this hack
with the roam D-Bus command.

BUG=b:159468518
TEST=network_WiFi_RoamFT.*, check logs to make sure FT is being used

Cq-Depend: chromium:2336194
Change-Id: I62d19e74a41efd0f1d3cd56c16ba5d01a24806b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2317111
Tested-by: Brian Norris <briannorris@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>