autotest: Check if FS is not read-only before start collecting the logs.
After analysis found that if linux switch FS to read-only then we
timeout on each file which taking time and we cannot start repair to fix
the DUT.

Collect logs is not priority  of repair task and if we have detected issue with filesystem we want to run verifier to check what is going and fix it.

Change-Id: I090def97fa74fa2bbf6f3d9b774212acedf52f36
BUG=b:161841773
TEST=running local

when FS is read-only and after.
./server/autoserv -s --host-info-subdir host_info_store -m   chromeos6-row6-rack15-host8  --lab True --local-only-host-info True -R -r /tr/
./server/autoserv -s --host-info-subdir host_info_store -m   chromeos6-row9-rack5-host1  --lab True --local-only-host-info True -R -r /tr/
./server/autoserv -s --host-info-subdir host_info_store -m   chromeos6-row8-rack8-host8  --lab True --local-only-host-info True -R -r /tr/
when system is read-only or not available it will spit attempt to collect logs.
When system is good to read then collecting logs is in process.

Change-Id: Id83f7eff03c3a303f9c5c66a13c7ab88be22a8cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2311317
Commit-Queue: Otabek Kasimov <otabek@google.com>
Tested-by: Otabek Kasimov <otabek@google.com>
Auto-Submit: Otabek Kasimov <otabek@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>