autotest: Avoid using ping to determine DUT is down when using port forwarding
If the host and DUT are on different networks, the DUT is likely exposed
to the host using SSH port forwarding. In this situation, `ping` is not
a useful measure of whether the DUT is up/down since a ping to localhost
will always succeed. Instead, fall back to the SSH way of doing things.

BUG=None
TEST=test_that localhost:2222 --args "chameleon_host=localhost \
       chameleon_ssh_port=2223" display_Resolution.reboot

Change-Id: I06bb34a2ab3dcea764c3de027bdafd9f75ad2e2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2331113
Reviewed-by: Otabek Kasimov <otabek@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Tested-by: Anand K Mistry <amistry@chromium.org>
Commit-Queue: Anand K Mistry <amistry@chromium.org>