network/iw_event_logger: don't greedily match colons (":")
The current parsing turns `iw event` messages like this:

  1596503812.227582: wlan0 (phy #0): disconnected (local request) reason: 1: Unspecified

into entry.message fields like this:

  Unspecified

This is because the usage of '.*' matches the longest possible match, as
long as the rest of the regex still matches.

This unbreaks _check_message_for_disconnect() (and therefore
get_disconnect_count(), etc.), which currently doesn't know how to find
the "disconnected" in the above log message.

BUG=b:159468518
TEST=`network_WiFi_RoamFT.*`, with new CL using
     assert_disconnect_count(1)

Change-Id: Idff6c313355badef270c1ade8ed6d9994904c09f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2336194
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>