firmware_TPMExtend: Check for firmware version and skip parts if too old
The test verifies that HWID is measured into the TPM's PCR1, but that
was only added in CL:226408. For older firmware versions the test fails
because that part simply isn't done (to no discernible disadvantage).

Since that commit was merged in 6425.0.0, check if the firmware is
older than that and, if so, skip the HWID part of the test.

Side note: There have been two cherry-picks into older firmware
branches (samus and storm), but since there are other firmware branches
without such a cherry-pick, just stick to 6425 as the cut-off.

BUG=b:159940406
TEST=Messed with the reported number a bit and verified that all
paths are executed as intended.

Change-Id: If254f8ba1117d08f8bb19594464fa28f221adfdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2339254
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>