WiFi_BSSTMReq: Test both "Scan" and "Probe" cases
When a DUT receives a BSS Transition Management Request it can behave in
two different ways. If the destination BSSID was already in a previous
scan result, the DUT directly connects to the destination BSSID since it
knows about it already. If the destination BSSID is not already known to
the DUT, the DUT will send a Probe Request specifically to that BSSID
and expect a Probe Response.

This CL adds the second case (Probe Request/Response) to the test.

BUG=b:160200843
TEST=test_that --iterations=100 ${HATCH_DUT} network_WiFi_BSSTMReq, only
2 failures not directly related to the change (device never saw the
Transition Request)
TEST=test_that --iterations=5 ${ZORK_DUT} network_WiFi_BSSTMReq

Change-Id: If3a2b17d285f7ab2f35a8754e2c92a9e8ca255ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2342587
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>