Remove provision_AutoUpdate.double from testing pool
Provision_AutoUpdate.double was being used in the past to make sure
changes to AU won't break provisioning flow in the lab. Basically to
make sure an image can update to itself. But since we don't use AU for
provisioning anymore, we don't need this test. We can add a simplified
version of this later that uses AU process in autotests rather than
chromite.

BUG=chromium:1113430
TEST=None

Change-Id: I0ea47d2a3f383f90aa6753278d44deed448af586
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2340246
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>